+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add subscription button to donation widget.
+- Added SCSS variable for background image for easier customization.
 
 ### Changed